# src/Routers/Docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IProfileI": {
				"properties": {
					"id": {
						"type": "string"
					},
					"UserId": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"UserId",
					"firstName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfile": {
				"properties": {
					"UserId": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"UserId",
					"firstName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateProfile": {
				"properties": {
					"id": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"firstName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IId": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGENjwt": {
				"properties": {
					"statusCode": {
						"type": "string"
					},
					"payload": {
						"properties": {
							"data": {
								"anyOf": [
									{},
									{
										"properties": {
											"message": {
												"type": "string"
											},
											"token": {
												"type": "string"
											}
										},
										"type": "object"
									}
								]
							}
						},
						"type": "object"
					}
				},
				"required": [
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IToken": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRecordOfAny": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			},
			"IUserI": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateUser": {
				"properties": {
					"username": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"salt": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"id",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "fastify",
		"version": "1.0.0",
		"description": "nodejs with fastify",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Morteza Ahmadi"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/profile": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IProfileI"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfile"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateProfile"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IId"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Patch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/profile/id/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/test": {
			"get": {
				"operationId": "GetTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"test": {
											"type": "string"
										}
									},
									"required": [
										"test"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "TestPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"test": {
											"type": "string"
										}
									},
									"required": [
										"test"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "TestUpdate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"test": {
											"type": "string"
										}
									},
									"required": [
										"test"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Testdelete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"test": {
											"type": "string"
										}
									},
									"required": [
										"test"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/test/id/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"test": {
											"type": "string"
										}
									},
									"required": [
										"test"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/test/jwt-generate": {
			"get": {
				"operationId": "JwtGenerate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGENjwt"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/test/jwt-decorate": {
			"post": {
				"operationId": "JwtDecorate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGENjwt"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IToken"
							}
						}
					}
				}
			}
		},
		"/test/jwt-check": {
			"get": {
				"operationId": "JwtCheck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGENjwt"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/test/redis/{key}": {
			"get": {
				"operationId": "GetRedis",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGENjwt"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/test/redis": {
			"post": {
				"operationId": "PostRedis",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGENjwt"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRecordOfAny"
							}
						}
					}
				}
			}
		},
		"/test/rabbitmq": {
			"post": {
				"operationId": "Rabbitmq",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGENjwt"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRecordOfAny"
							}
						}
					}
				}
			}
		},
		"/user": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IUserI"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserI"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserI"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateUser"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IId"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Patch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/id/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserI"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"token"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}